requirements.txt
llama-index
chromadb
torch
rich
python-dotenv
transformers
huggingface-hub
accelerate
# Optional: Groq LLM integration
groq
llama-index-vector-stores-chroma
llama-index-llms-groq
llama-index-embeddings-huggingface
